+++ src/main/java/com/raoulvdberge/rangedpumps/proxy/ProxyClient.java
@@ -1,12 +1,12 @@
-package rangedpumps.proxy;
+package com.raoulvdberge.rangedpumps.proxy;
 
+import com.raoulvdberge.rangedpumps.RangedPumps;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.item.Item;
 import net.minecraftforge.client.model.ModelLoader;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import rangedpumps.RangedPumps;
 
-public class ClientProxy extends CommonProxy {
+public class ProxyClient extends ProxyCommon {
     @Override
     public void preInit(FMLPreInitializationEvent e) {
         super.preInit(e);