+++ src/main/java/rangedpumps/proxy/ServerProxy.java
@@ -1,4 +1,0 @@
-package rangedpumps.proxy;
-
-public class ServerProxy extends CommonProxy {
-}